Allow passing in a root scope for reordering.
If a model only makes sense within the context of the parent model it
belongs to, then positioning should reflect that. If we had a parent
'Event' model that has many 'Locations', then after running a rebuild
our positions would look like this:

```
Event A
- 0 | Location A
- 1 | Location B
  - 0 | Location B.1

Event B
- 2 | Location A
- 3 | Location B
  - 0 | Location B.1
```

If we add a scope to the reordering SQL statement, we can speed up
reordering, and give us positions numbers based on the parent object,
and not the base child table.

With these changes our positions would like this after a rebuild:

```
Event A
- 0 | Location A
- 1 | Location B
  - 0 | Location B.1

Event B
- 0 | Location A
- 1 | Location B
  - 0 | Location B.1
```